write_graphite plugin: Decrese buffer size.
The larger the buffer, the longer you need to wait for the data to be written
to Carbon. The new buffer size assumes an MTU of 1500 bytes (Ethernet payload)
and keeps margins for IPv6 and TCP (with timestamps). This should decrease the
number of packets needed (because data is now cleanly aligned to the packet
size) and decrease the overall time spent buffered.

Change-Id: I73021608f1ff0f22f9e73cd62b7fb5ea79421546

Signed-off-by: Scott Sanders <scott@jssjr.com>